made upgrades much easier to understand. More of the same type are not less effective, but instead they can consume more ink to run.

+++ src/main/java/de/dafuqs/spectrum/blocks/upgrade/UpgradeBlock.java
@@ -1,27 +1,22 @@
 package de.dafuqs.spectrum.blocks.upgrade;
 
-import de.dafuqs.spectrum.networking.SpectrumS2CPacketSender;
-import de.dafuqs.spectrum.particle.SpectrumParticleTypes;
-import de.dafuqs.spectrum.particle.effect.ColoredTransmission;
-import de.dafuqs.spectrum.registries.SpectrumSoundEvents;
+import de.dafuqs.spectrum.networking.*;
+import de.dafuqs.spectrum.particle.*;
+import de.dafuqs.spectrum.particle.effect.*;
+import de.dafuqs.spectrum.registries.*;
 import net.minecraft.block.*;
-import net.minecraft.block.entity.BlockEntity;
-import net.minecraft.entity.ai.pathing.NavigationType;
-import net.minecraft.server.world.ServerWorld;
-import net.minecraft.sound.SoundCategory;
-import net.minecraft.util.DyeColor;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.Vec3d;
-import net.minecraft.util.math.Vec3i;
-import net.minecraft.util.shape.VoxelShape;
-import net.minecraft.world.BlockView;
-import net.minecraft.world.World;
-import net.minecraft.world.event.BlockPositionSource;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
+import net.minecraft.block.entity.*;
+import net.minecraft.entity.ai.pathing.*;
+import net.minecraft.server.world.*;
+import net.minecraft.sound.*;
+import net.minecraft.util.*;
+import net.minecraft.util.math.*;
+import net.minecraft.util.shape.*;
+import net.minecraft.world.*;
+import net.minecraft.world.event.*;
+import org.jetbrains.annotations.*;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.*;
 
 public class UpgradeBlock extends BlockWithEntity {
 
@@ -66,10 +61,10 @@ public class UpgradeBlock extends BlockWithEntity {
 	// Like: The further the player progresses,
 	// the higher are the chances for good mods?
 	private final Upgradeable.UpgradeType upgradeType;
-	private final float upgradeMod;
+	private final int upgradeMod;
 	private final DyeColor effectColor;
 
-	public UpgradeBlock(Settings settings, Upgradeable.UpgradeType upgradeType, float upgradeMod, DyeColor effectColor) {
+	public UpgradeBlock(Settings settings, Upgradeable.UpgradeType upgradeType, int upgradeMod, DyeColor effectColor) {
 		super(settings);
 		this.upgradeType = upgradeType;
 		this.upgradeMod = upgradeMod;
@@ -146,7 +141,7 @@ public class UpgradeBlock extends BlockWithEntity {
 		return this.upgradeType;
 	}
 
-	public float getUpgradeMod() {
+	public int getUpgradeMod() {
 		return this.upgradeMod;
 	}
